pyproject.toml
[project]
name = "gridspace-guava"
version = "0.18.0"
description = "The official Python SDK for building Guava voice agents."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.55.0",
    "openai>=2.8.1",
    "pydantic>=2.12.4",
    "typing-extensions>=4.15.0",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
lancedb = ["lancedb>=0.20.0"]
pgvector = ["psycopg[binary]>=3.2.0", "pgvector>=0.4.0", "numpy>=1.26.0"]
chromadb = ["chromadb>=0.6.0"]
pinecone = ["pinecone>=5.0.0"]

[dependency-groups]
dev = [
    "aiortc>=1.14.0",
    "fastapi>=0.128.0",
    "mypy>=1.18.2",
    "numpy>=2.2.6",
    "ruff>=0.14.6",
    "semgrep>=1.152.0",
    "simple503>=0.4.0",
    "sounddevice>=0.5.5",
    "ty>=0.0.21",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["guava"]
exclude = ["internal"]

[tool.hatch.build.targets.sdist]
exclude = ["internal"]
only-include = ["guava", "pyproject.toml", "README.md"]

[tool.ruff]
line-length = 100
